pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "monodepth2-kitti"
version = "0.1.0"
description = "Monocular depth estimation with Monodepth2 on KITTI (PyTorch + TensorFlow)."
authors = [{ name = "Open Source Contributors" }]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["monodepth2", "kitti", "depth-estimation", "pytorch", "tensorflow"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "tensorflow>=2.12.0",
  "mindspore==2.2.10",
  "numpy>=1.24.0",
  "scipy>=1.10.0",
  "matplotlib>=3.7.0",
  "opencv-python>=4.8.0",
  "Pillow>=10.0.0",
  "pyyaml>=6.0",
  "tqdm>=4.66.0",
  "rich>=13.7.0",
  "pytest>=7.4.0",
  "pytest-mock>=3.10.0",
]

[project.optional-dependencies]
dev = ["black", "isort", "flake8", "mypy"]

[tool.setuptools.packages.find]
where = ["src", "models"]
